backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
websockets==12.0
langchain==0.0.350
langgraph==0.0.20
openai==1.3.7
anthropic==0.7.7
pgvector==0.2.4
alembic==1.13.1
httpx==0.25.2
python-dotenv==1.0.0
boto3==1.34.0
pillow==10.1.0
numpy==1.25.2
torch==2.1.1
torchvision==0.16.1
trimesh==4.0.5
open3d==0.17.0
xatlas==0.0.1
ffmpeg-python==0.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
